requirements.txt
Django~=4.1.9
django-ckeditor~=6.5.1
django-filter~=22.1
APScheduler~=3.10.1